requirements.txt
numpy>=1.17
scikit-image<0.17
scipy
tqdm
pillow
